+++ scripts/macros/2024/classFeatures/bard/dance/tandemFootwork.js
@@ -35,7 +35,7 @@ export let tandemFootwork = {
             }
         ]
     },
-    createItem: [
+    item: [
         {
             pass: 'created',
             macro: added