+++ scripts/macros/2024/classFeatures/bard/glamour/mantleOfInspiration.js
@@ -55,7 +55,7 @@ export let mantleOfInspiration = {
             }
         ]
     },
-    createItem: [
+    item: [
         {
             pass: 'created',
             macro: added