+++ scripts/macros/2024/classFeatures/cleric/life/preserveLife.js
@@ -54,7 +54,7 @@ export let preserveLife = {
             }
         ]
     },
-    createItem: [
+    item: [
         {
             pass: 'created',
             macro: added