+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - Interception
 ## Bug Fixes
 - Grease
+## API Updates:
+- Added the an "Item" event, which can run macros when an item is created, deleted, equipped, or unequipped. The "Create Item" event has been replaced with this.
 
 # 1.2.35 Release:
 ## Bug Fixes: